refactor(typescript): typescript declaration files
