+++ test/fixtures/apps/app-ts/app/proxy/foo.d.ts
@@ -1,3 +1,6 @@
+// NOTE
+import 'egg';
+
 declare module 'egg' {
   interface Foo {
     bar(): string;
