+++ test/fixtures/apps/app-ts/tsconfig.json
@@ -4,7 +4,7 @@
     "baseUrl": ".",
     "paths": {
       "egg": [
-        "../../../../index"
+        "../../../.."
       ]
     },
     "types" : ["node"],
