+++ types/tsconfig.json
@@ -1,0 +1,21 @@
+{
+  "compilerOptions": {
+    "strict": true,
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "noEmit": true,
+    "lib": ["es2017"],
+    "esModuleInterop": true,
+
+    "baseUrl": ".",
+    "paths": {
+      "egg": [
+        ".."
+      ],
+      "egg/*": [
+        "../"
+      ]
+    }
+  },
+  "types": ["node"]
+}
